feat: Update reminder duration handling to use Int instead of Long, add MinutesSelect component for custom duration input

+++ app/src/main/java/com/github/meypod/al_azan/core/domain/model/reminder/Reminder.kt
@@ -17,7 +17,8 @@ data class Reminder(
     val label: String = "",
     val enabled: Boolean = false,
     val prayer: Prayer,
-    val duration: Long,
+    /** offset from the prayer time, in minutes */
+    val duration: Int,
     val durationModifier: Int,
     val sound: ReminderAudioEntry? = null,
     val once: Boolean? = null,
